Support F1 accelerator even when the focus is in CommandBox and ResultDisplay
This issue may be caused by a JavaFX bug, as suggested here
https://bugs.openjdk.java.net/browse/JDK-8131666.

It's because TextField and TextAreas in CommandBox and ResultDisplay
will consume F1 keyevents. As a result, the accelerators will not be
fired up.

The discussion in the post shows that the bug will only be fixed in
later version of JDK (though it's openJDK but I guess it should be
similar case for JDK itself.)

Thus what I implemented in this commit is just a workaround.