logic: call Storage#saveAddressBook(...) directly
When the model's address book is modified, `ModelManager` will raise an
`AddressBookChangedEvent` via `EventsCenter`. `StorageManager` will then
respond by calling `StorageManager#saveAddressBook(...)` to save the
address book.

If an error occurs while saving the address book, `StorageManager` will
raise a `DataSavingExceptionEvent` via `EventsCenter`. `UiManager` will
then respond by showing an alert dialog.

However, we are trying to move away from EventsCenter. Furthermore, all
of these "communication by events" is unnecessary. Let's just let
`LogicManager` call `Storage#saveAddressBook(...)` directly. If an error
occurs, this can then be sent to the caller via exceptions. The end
result is that the control flow is much clearer and explicit in the
code.

Note that the new implementation differs in the handling of address book
storage errors: previously an error dialog would be shown. Now, the
error would be reported by the ResultDisplay. This approach is chosen as
its much simpler -- we can simply reuse the same battle-tested code
paths for handling command errors. Furthermore, the overall goal is
still achieved as the error is still reported to the user.