MainApp: don't forcefully terminate the application on exit
When MainApp receives an ExitAppRequestEvent, it forcefully terminates
the application with:

    Platform.exit();
    System.exit(0);

However, this makes testing the `exit` command extremely hard. This is
because in the JUnit environment, calling System.exit(...) will cause
the test runner to terminate as well, causing tests to not properly be
run.

But really, we don't need to call Platform.exit() or System.exit().
This is because by default, once all windows are closed, JavaFX will
automatically begin the shutdown process by calling Application#stop().

As such, let's introduce one very important change:

  * Rename Ui#stop() to Ui#hide(), to better describe what it does and
    to make it clear that it could be called multiple times.

And then rewrite the application `exit` process as follows:

  * When MainApp receives an ExitAppRequestEvent, it calls Ui#hide().

  * UiManager#hide() will call MainWindow#hide().

  * MainWindow#hide() will hide _both_ itself and HelpWindow. This
    ensures that all windows in the application are closed.

  * Finally, JavaFX will notice that all windows are closed, and call
    MainApp#stop().

  * MainApp#stop() will call UiManager#hide() again. This is because it
    may have been called because of Platform.exit().

  * MainApp#stop() would then perform the rest of the shutdown sequence
    (e.g. writing the userprefs file).